refactor: consolidate duplicated utility modules
- Consolidate duplicate utils modules into centralized locations
- Maintain backward compatibility with re-export files
- Improve documentation and module organization
- Remove debug file from source control
